Fix bug : Display connect fails when app is switched to foreground
:Release Notes:
Fix bug : Display connect fails when app is switched to foreground

:Detailed Notes:
There is a problem that the MDC can not connect to the video when the
video client app repeats the background and the foreground.
If the app enters the foreground, "TryDisplay" connect should send an
event post to the audio / video element, but there is a problem sending
a TryDisplay event to the audio only.
Modified to send TryDisplay event in video element also.

:Testing Performed:
1. Run Browser
2. Input "http://camendesign.com/code/video_for_everybody/test.html
in URL bar
3. Play video
4. Open new tab
5. Input "www.google.com"
6. switch to first tab
7. switch to second tab
8. repeat 6,7 2times quickly.
9. First tab should show video at the correct position even if repeat
step 8

:QA Notes:
N/A

:Issues Addressed:

Change-Id: Id6d37cff84f961a2221231dffa11a1088c7684d0