Fix wrong value for is_managed for an unmanaged pipeline
:Release Notes:
Fix wrong value for is_managed for an unmanaged pipeline

:Detailed Notes:
Even if we run unmanaged pipeline, the is_managed value is shown true
from ums-status.
Unmanaged pipeline is not registered to pipeline manager.
So before calling rm->setManaged, we have to check whether it is managed
or unmanged pipeline.

:Testing Performed:
To verify proper is_managed value is set using ums-status

:QA Notes:

:Issues Addressed:

Change-Id: I553781dbbeb3d3dab9a8bcda133a160c7c9bb805