Fix SetDisplayWindow error for secondary display
:Release Notes:
Fix SetDisplayWindow error for
secondary display

:Detailed Notes:
Fix an error in which SetDisplayWindow
was not applied to secondary display.
Since umediaserver always try to connect
to primary display first, It has been
changed to connect to acquired display
resource.

:Testing Performed:
Launch Youtube Application in secondary display
and check whether a media content is played with
properly scaled size.

:QA Notes:

:Issues Addressed:

Change-Id: Ica02504d5c9a34875789027267d52ec1921fd2d5