Fix the static issues in umediaserver component
:Release Notes:
Addressed the dereferencing of null pointer and memory management issues
for umediaserver

:Detailed Notes:
Added the code to resolve dereferencing of null pointer and Fixed a
memory management issue by removing a raw pointer to be assigned to
a shared pointer. As a solution,Converted the raw pointer to
smart pointer.

:Testing Performed:
QWTC-474,QWTC-475,QWTC-476,QWTC-477,QWTC-492,QWTC-496

:QA Notes:
N/A

:Issues Addressed:
[WRP-10762] Fix coverity issues in umediaserver to avoid null
dereferencing

Change-Id: I1fa27a2aad1c151f5d4b919ef8fd311c3e51752c